US28: Add raw data to the new writer.
- The new writer writes all the required tags and the raw data. No previews yet.
- Fixed orientation value in MetaData.

+++ MetaData.hpp
@@ -37,7 +37,7 @@ public:
     bool isSameFormat(const MetaData & r) {
         return width == r.width && height == r.height && filters == r.filters && cdesc == r.cdesc;
     }
-    int FC(int row, int col) const {
+    uint8_t FC(int row, int col) const {
         return (filters >> (((row << 1 & 14) | (col & 1)) << 1) & 3);
     }
     double logExp() const;
